refactor: split provider display info and share fallback metrics
