+++ docs/CODEBASE_AUDIT_ACTION_TABLE_2026-03-09.md
@@ -59,14 +59,15 @@ This table captures every issue found in this pass. It is broad and high-signal,
 | R39 | Fixed | Usage-view materialization split | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_materialize.go` | Temp-table creation/indexing/cleanup and aggregate initialization moved out of the main usage-view orchestration path. | Continue splitting aggregate query fanout if `usage_view.go` grows again. |
 | R40 | Fixed | Analytics cost fallback extraction | `internal/core/analytics_costs.go`, `internal/tui/analytics_data.go` | Analytics all-time/today/week cost fallback rules now live in shared core logic instead of TUI-owned metric-key decoding. | Continue moving remaining analytics/detail metric decoding into shared extractors. |
 | R41 | Fixed | Usage-view aggregate fanout split | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_aggregate.go` | Query fanout and aggregate assembly now live in a dedicated helper instead of inline in the main usage-view orchestration path. | Continue splitting only if the aggregate helper grows materially. |
+| R42 | Fixed | Provider display-info split and shared fallback metric helpers | `internal/tui/model.go`, `internal/tui/model_display_info.go`, `internal/core/dashboard_display_metrics.go` | Provider tile display-summary logic moved out of the main TUI model file, and fallback/rate-limit metric selection now lives in shared core helpers instead of ad hoc TUI parsing. | Continue moving the remaining analytics/detail-specific metric decoding into shared extractors. |
 
 ## Action Table
 
 | ID | Priority | Area | Evidence | Issue | Recommended action | Expected payoff |
 | --- | --- | --- | --- | --- | --- | --- |
 | A1 | P2 | Account config contract hardening | `internal/core/provider.go:31-43`, `internal/config/config.go:199-206` | Path overload dependence is removed from the hot runtime flow, but `Binary` / `BaseURL` still coexist in the same type and the distinction between CLI path vs provider-local path is still not encoded by type. | Introduce a dedicated typed runtime-hints/path struct and eventually retire path-related legacy comments/compatibility in `AccountConfig`. | Finishes the contract cleanup and makes misuse harder. |
-| A2 | P2 | TUI/application decomposition follow-through | `internal/tui/model.go`, `internal/tui/detail.go`, `internal/tui/detail_tokens.go`, `internal/dashboardapp/service.go` | Side effects are injected and some detail logic is split, but TUI state-transition and rendering logic is still concentrated in very large files. | Continue decomposing render-heavy/detail/settings flows and move more orchestration decisions into the dashboard application layer over time. | Lower UI complexity and smaller blast radius per change. |
-| A3 | P2 | UI metric-prefix parsing | `internal/tui/analytics.go`, `internal/tui/detail.go`, `internal/core/usage_breakdowns.go`, `internal/core/analytics_snapshot.go`, `internal/core/analytics_costs.go` | Composition bars, analytics model views, and analytics cost fallback now consume shared extractors, but some analytics/detail sections still decode metric-key conventions directly in TUI code. | Continue promoting the remaining analytics/detail extractors into `internal/core` or `internal/telemetry` so renderers consume structured sections instead of re-parsing maps. | Removes a large class of UI drift bugs and reduces per-render work. |
+| A2 | P2 | TUI/application decomposition follow-through | `internal/tui/model.go`, `internal/tui/model_display_info.go`, `internal/tui/detail.go`, `internal/tui/detail_tokens.go`, `internal/dashboardapp/service.go` | Side effects are injected, provider display-info logic is split out, and more detail logic is isolated, but TUI state-transition and rendering logic are still concentrated in very large files. | Continue decomposing render-heavy/detail/settings flows and move more orchestration decisions into the dashboard application layer over time. | Lower UI complexity and smaller blast radius per change. |
+| A3 | P2 | UI metric-prefix parsing | `internal/tui/analytics.go`, `internal/tui/detail.go`, `internal/core/usage_breakdowns.go`, `internal/core/analytics_snapshot.go`, `internal/core/analytics_costs.go`, `internal/core/dashboard_display_metrics.go` | Composition bars, provider tile fallback/rate-limit selection, analytics model views, and analytics cost fallback now consume shared extractors, but some analytics/detail sections still decode metric-key conventions directly in TUI code. | Continue promoting the remaining analytics/detail extractors into `internal/core` or `internal/telemetry` so renderers consume structured sections instead of re-parsing maps. | Removes a large class of UI drift bugs and reduces per-render work. |
 | A4 | P2 | Large provider monolith follow-through | `internal/providers/ollama/ollama.go`, `internal/providers/zai/zai.go`, `internal/providers/gemini_cli/gemini_cli.go`, `internal/providers/copilot/copilot.go`, `internal/providers/claude_code/claude_code.go`, `internal/providers/codex/codex.go` | Cursor and OpenRouter are now materially decomposed, but several other providers still combine transport, parsing, normalization, and projection in single 1900-2600 LOC files. | Split the remaining large providers by concern: account/API fetch, local-data adapters, projection helpers, and telemetry helpers. | Smaller diffs, less drift risk, and easier provider-specific testing. |
 | A6 | P2 | Telemetry usage-view monolith | `internal/telemetry/usage_view.go`, `internal/telemetry/usage_view_helpers.go`, `internal/telemetry/usage_view_projection.go`, `internal/telemetry/usage_view_queries.go`, `internal/telemetry/usage_view_materialize.go`, `internal/telemetry/usage_view_aggregate.go` | The usage-view code is materially smaller after the helper/projection/query/materialization/aggregate splits, but the top-level orchestration path still coordinates caching, source selection, and final snapshot application in one place. | Continue splitting only if future telemetry work reintroduces sprawl, and consider a typed intermediate aggregation model if query optimization pressure grows. | Easier optimization and safer incremental changes. |
 | A8 | P3 | Ambiguous local-source account attribution still requires explicit disambiguation | `internal/daemon/source_collectors.go`, `internal/daemon/server_http.go`, `cmd/openusage/telemetry.go` | Unambiguous local collectors now bind to configured accounts, but when multiple accounts share the same source paths the daemon intentionally falls back to source-scoped attribution rather than guessing. This is correct, but it still leaves ambiguous setups dependent on explicit account selection. | If multi-account local-source workflows become common, add persisted per-source alias mapping or require explicit source/account binding in config for ambiguous path groups. | Makes the remaining ambiguity explicit instead of silent, and defines the next hardening step only if needed. |
